Add suport for Markdown formatting

+++ app/Http/Controllers/IdeaImageController.php
@@ -12,9 +12,11 @@ class IdeaImageController extends Controller
     {
         Gate::authorize('workWith', $idea);
 
-        Storage::disk('public')->delete($idea->image_path);
+        if ($idea->image_path) {
+            Storage::disk('public')->delete($idea->image_path);
 
-        $idea->update(['image_path' => null]);
+            $idea->update(['image_path' => null]);
+        }
 
         return back();
     }
